Summarize continuous and discrete variables.

A vector based summary function. Provides a table with percentages for discrete data. And median, mean, range or standard deviation for continuous data.

get.summary(var,type, var.n=NULL, type2="range")

Arguments

var
A vector, can be discrete(integer, character or factor) or continuous(numeric)

type
integer

type2
if type=2 specify whether you want range or standard deviation. default is - "range", For SD pass "sd"

var.n
character. specify the name of the variable that you are summarizing. Default is NULL

Details

Summary statistics are rounded to nearest 2 decimal points.

Value

if type=1 returns a character vector with summary information. length of the vector is equal to length of unique classes and NA if any.

if type=2 returns a character of length 1 with summary information. NAs are separated by ";"

See also

Examples

#when type=1, data is discrete
set.seed(123)
x<-sample(1:3,50, replace=TRUE)
get.summary(x,type=1)

#>    [,1]     
#> =1 "16(32%)"
#> 2  "15(30%)"
#> 3  "19(38%)"

#lets see with NAs
x[c(1,5,30)] = NA
get.summary(x,type=1)

#>      [,1]     
#> =1   "16(32%)"
#> 2    "14(28%)"
#> 3    "17(34%)"
#> <NA> "3(6%)"  

#when type=2, data is continuous
set.seed(123)
x<-rnorm(50)
get.summary(x, type=2, type2="range")

#>      [,1]                     
#> [1,] "-0.07(-1.97,2.17); NA=0"

x[c(1,5,30)] = NA
get.summary(x, type=2, type2="sd")

#>      [,1]               
#> [1,] "-0.08(0.93); NA=3"